pyproject.toml
[project]
    name = "FileFinder9000"
    version = "0.1.1"
    description = "Perform bulk searches for files using Everything csv exports."
    readme = "README.md"
    authors = [{ name = "Nathan Vaughn", email = "nath@nvaughn.email" }]
    classifiers = [
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Programming Language :: Python :: 3.14",
    ]
    requires-python = ">=3.9"
    dependencies = []

[dependency-groups]
    dev = ["pillow>=11.3.0", "pre-commit>=3.2.0", "pyinstaller>=6.17.0"]

[project.urls]
    Homepage   = "https://github.com/NathanVaughn/FileFinder9000"
    Repository = "https://github.com/NathanVaughn/FileFinder9000.git"
    Issues     = "https://github.com/NathanVaughn/FileFinder9000/issues"

[tool.pyright]
    typeCheckingMode           = "standard"
    venvPath                   = "."
    venv                       = ".venv"
    reportMissingParameterType = true

[tool.ruff.lint.flake8-tidy-imports]
    ban-relative-imports = "all"
